Directory restructuring.
